+++ sdk/storage/storage-blob/src/ServiceURL.ts
@@ -1,6 +1,8 @@
+import { HttpResponse } from "@azure/ms-rest-js";
+
 import * as Models from "../src/generated/src/models";
 import { Aborter } from "./Aborter";
-import { ListContainersIncludeType } from "./generated/src/models/index";
+import { ListContainersIncludeType } from "./generated/src/models";
 import { Service } from "./generated/src/operations";
 import { Pipeline } from "./Pipeline";
 import { StorageURL } from "./StorageURL";
@@ -28,6 +30,81 @@ export interface IServiceListContainersSegmentOptions {
    */
   include?: ListContainersIncludeType;
 }
+
+export interface UserDelegationKey {
+  /**
+   * The Azure Active Directory object ID in GUID format.
+   *
+   * @type {string}
+   * @memberof UserDelegationKey
+   */
+  signedOid: string;
+  /**
+   * The Azure Active Directory tenant ID in GUID format
+   *
+   * @type {string}
+   * @memberof UserDelegationKey
+   */
+  signedTid: string;
+  /**
+   * The date-time the key is active
+   *
+   * @type {Date}
+   * @memberof UserDelegationKey
+   */
+  signedStart: Date;
+  /**
+   * The date-time the key expires
+   *
+   * @type {Date}
+   * @memberof UserDelegationKey
+   */
+  signedExpiry: Date;
+  /**
+   * Abbreviation of the Azure Storage service that accepts the key.
+   *
+   * @type {string}
+   * @memberof UserDelegationKey
+   */
+  signedService: string;
+  /**
+   * The service version that created the key.
+   *
+   * @type {string}
+   * @memberof UserDelegationKey
+   */
+  signedVersion: string;
+  /**
+   * The key as a base64 string.
+   *
+   * @type {string}
+   * @memberof UserDelegationKey
+   */
+  value: string;
+}
+
+export declare type ServiceGetUserDelegationKeyResponse = UserDelegationKey &
+  Models.ServiceGetUserDelegationKeyHeaders & {
+    /**
+     * The underlying HTTP response.
+     */
+    _response: HttpResponse & {
+      /**
+       * The parsed HTTP response headers.
+       */
+      parsedHeaders: Models.ServiceGetUserDelegationKeyHeaders;
+
+      /**
+       * The response body as text (string format)
+       */
+      bodyAsText: string;
+
+      /**
+       * The response body as parsed JSON or XML
+       */
+      parsedBody: Models.UserDelegationKey;
+    };
+  };
 
 /**
  * A ServiceURL represents a URL to the Azure Storage Blob service allowing you
@@ -172,5 +249,53 @@ export class ServiceURL extends StorageURL {
       marker,
       ...options
     });
+  }
+
+  /**
+   * ONLY AVAILABLE WHEN USING BEARER TOKEN AUTHENTICATION (TokenCredential).
+   *
+   * Retrieves a user delegation key for the Blob service. This is only a valid operation when using
+   * bearer token authentication.
+   *
+   * @see https://docs.microsoft.com/en-us/rest/api/storageservices/get-user-delegation-key
+   *
+   * @param {Aborter} aborter Create a new Aborter instance with Aborter.none or Aborter.timeout(),
+   *                          goto documents of Aborter for more examples about request cancellation
+   * @param {Date} start      The start time for the user delegation SAS. Must be within 7 days of the current time
+   * @param {Date} expiry     The end time for the user delegation SAS. Must be within 7 days of the current time
+   * @returns {Promise<ServiceGetUserDelegationKeyResponse>}
+   * @memberof ServiceURL
+   */
+  public async getUserDelegationKey(
+    aborter: Aborter,
+    start: Date,
+    expiry: Date
+  ): Promise<ServiceGetUserDelegationKeyResponse> {
+    const response = await this.serviceContext.getUserDelegationKey(
+      {
+        start: start.toISOString(),
+        expiry: expiry.toISOString()
+      },
+      {
+        abortSignal: aborter
+      }
+    );
+
+    const res: ServiceGetUserDelegationKeyResponse = {
+      _response: response._response,
+      requestId: response.requestId,
+      version: response.version,
+      date: response.date,
+      errorCode: response.errorCode,
+      signedOid: response.signedOid,
+      signedTid: response.signedTid,
+      signedStart: new Date(response.signedStart),
+      signedExpiry: new Date(response.signedExpiry),
+      signedService: response.signedService,
+      signedVersion: response.signedVersion,
+      value: response.value
+    };
+
+    return res;
   }
 }